requirements.txt
# Django & REST
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3

# Database
psycopg2-binary>=2.9

# Async Tasks
celery>=5.3
redis>=5.0

# Server
gunicorn>=21.2

# Image handling
Pillow>=10.0
